refactor(agentic-tools): rename sandbox → runner to free name for substrate
Industry has converged on "sandbox" for capability-aware execution
environments (E2B, Anthropic computer use, Daytona, Modal, Cloudflare
Workers, OpenAI playground). The forthcoming pkg/sandbox substrate
primitive (see docs/proposals/sandbox-substrate.md) needs that name;
renaming the existing tool-level HTTP client now — before the
substrate lands — avoids perpetual "which sandbox?" disambiguation
across the codebase.

External operator-facing surfaces unchanged: SANDBOX_URL env var,
sandboxURL parameter name, remote-server wire protocol.

Internal renames:
- Package: agentic-tools/sandbox → agentic-tools/runner
- Field: e.sandbox *sandbox.Client → e.runner *runner.Client
- Method: execSandbox → execRemote
- Method: verifyCleanSandbox → verifyCleanRemote
- Test: TestBashExecutor_SandboxMode → TestBashExecutor_RemoteMode
